requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.136.0
beautifulsoup4==4.12.3